--- packages/venia-ui/lib/components/FilterModal/FilterFooter/index.js
@@ -1,1 +1,0 @@
-export { default } from './filterFooter';
