--- packages/venia-ui/lib/components/FilterModal/FilterSearch/index.js
@@ -1,1 +1,0 @@
-export { default } from './filterSearch';
