+++ packages/venia-ui/lib/components/FilterModal/filterSearch.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { useFormApi, useFieldState } from 'informed';
 import { Search as SearchIcon, X as ClearIcon } from 'react-feather';
 
-import { mergeClasses } from '../../../classify';
-import Icon from '../../Icon';
-import TextInput from '../../TextInput';
-import Trigger from '../../Trigger';
+import { mergeClasses } from '../../classify';
+import Icon from '../Icon';
+import TextInput from '../TextInput';
+import Trigger from '../Trigger';
 import defaultClasses from './filterSearch.css';
 
 const clearIcon = <Icon src={ClearIcon} size={18} />;
